Spawning: stop bothering to create a symlink for rom files with spaces, etc
This used to be necessary when the rom filename was passed directly to pnd_run.sh,
but for a while now emu apps are pre-mounted and their commands run directly,
making it possible to just quote the rom filename arg instead.

The symlinks caused problems when the emu expected to read other files from the
same path as the rom.